requirements.txt
flask==3.1.0
requests==2.32.3
playwright==1.49.1
openpyxl==3.1.5
waitress>=3.0.0
pyinstaller>=6.0,<7.0
pytest>=8.0
pytest-mock>=3.12
pyyaml>=6.0
